+++ Source/DotSpatial.Symbology/LayerFrame.cs
@@ -128,22 +128,19 @@ namespace DotSpatial.Symbology
         private void Configure()
         {
             Layers = new LayerCollection(this);
-            base.LegendText = SymbologyMessageStrings.LayerFrame_Map_Layers;
-            ContextMenuItems = new List<SymbologyMenuItem>();
-            ContextMenuItems.Add(new SymbologyMenuItem(SymbologyMessageStrings.LayerFrame_RemoveMapFrame, Remove_Click));
-            ContextMenuItems.Add(new SymbologyMenuItem(SymbologyMessageStrings.LayerFrame_ZoomToMapFrame, ZoomToMapFrame_Click));
-            ContextMenuItems.Add(new SymbologyMenuItem(SymbologyMessageStrings.LayerFrame_CreateGroup, CreateGroup_Click));
+            LegendText = SymbologyMessageStrings.LayerFrame_Map_Layers;
+            ContextMenuItems = new List<SymbologyMenuItem>
+            {
+                new SymbologyMenuItem(SymbologyMessageStrings.LayerFrame_ZoomToMapFrame, ZoomToMapFrame_Click),
+                new SymbologyMenuItem(SymbologyMessageStrings.LayerFrame_CreateGroup, CreateGroup_Click)
+            };
 
-            base.LegendSymbolMode = SymbolMode.GroupSymbol;
+            LegendSymbolMode = SymbolMode.GroupSymbol;
             LegendType = LegendType.Group;
             MapFrame = this;
             ParentGroup = this;
             _drawingLayers = new List<ILayer>();
         }
-
-        //private void Properties_Click(object sender, EventArgs e)
-        //{
-        //}
 
         private void CreateGroup_Click(object sender, EventArgs e)
         {
@@ -172,11 +169,6 @@ namespace DotSpatial.Symbology
                     OnExtentsChanged(_viewExtents);
                 }
             }
-        }
-
-        private void Remove_Click(object sender, EventArgs e)
-        {
-            OnRemoveItem();
         }
 
         private void ZoomToMapFrame_Click(object sender, EventArgs e)